docs: Add MIT License for Contributors
In the absence of anyone obvious to license this to, I have added a license for the Contributors.
I have defaulted to MIT, as it is easier to implement than GPL (which should be in every file) -
I am happy to change this as needed. For now, I figure this is a start.

What do you think?